Move Includes class to Facades namespace
